requirements.txt
tqdm==4.66.2
cupy-cuda12x==13.1.0
numpy==1.26.3
pandas==2.2.2
scikit-learn==1.5.0
scipy==1.13.1